rename folder pairedomics to strain
