+++ tests/pairedomics/test_strain_mappings_generator.py
@@ -1,12 +1,12 @@
 import json
-from nplinker.pairedomics import extract_mappings_ms_filename_spectrum_id
-from nplinker.pairedomics import extract_mappings_original_genome_id_resolved_genome_id
-from nplinker.pairedomics import extract_mappings_resolved_genome_id_bgc_id
-from nplinker.pairedomics import extract_mappings_strain_id_ms_filename
-from nplinker.pairedomics import extract_mappings_strain_id_original_genome_id
-from nplinker.pairedomics import get_mappings_strain_id_bgc_id
-from nplinker.pairedomics import get_mappings_strain_id_spectrum_id
-from nplinker.pairedomics import podp_generate_strain_mappings
+from nplinker.strain import extract_mappings_ms_filename_spectrum_id
+from nplinker.strain import extract_mappings_original_genome_id_resolved_genome_id
+from nplinker.strain import extract_mappings_resolved_genome_id_bgc_id
+from nplinker.strain import extract_mappings_strain_id_ms_filename
+from nplinker.strain import extract_mappings_strain_id_original_genome_id
+from nplinker.strain import get_mappings_strain_id_bgc_id
+from nplinker.strain import get_mappings_strain_id_spectrum_id
+from nplinker.strain import podp_generate_strain_mappings
 from nplinker.strain import Strain
 from nplinker.strain_collection import StrainCollection
